+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ master
   exactly during a SQL request on the invalidated table(s)
 - Uses infinite timeout
 - Adds a setting to choose if random queries must be cached
+- Caches all queries implying ``Queryset.extra``
 
 
 0.8.1